kalloc: fix issue of type casting

+++ kernel/kalloc.c
@@ -32,7 +32,7 @@ struct {
 
 int set_pageRefCnt(uint64 pa, uint8 cnt)
 {
-  if ((pa >= PHYSTOP) || (pa < end)) {
+  if ((pa >= PHYSTOP) || (pa < (uint64)end)) {
     return -1;
   }
   mem_map[IDX_MEM_MAP(pa)].refCnt = cnt;
@@ -42,7 +42,7 @@ int set_pageRefCnt(uint64 pa, uint8 cnt)
 int get_pageRefCnt(uint64 pa)
 {
   uint8 tmp;
-  if ((pa >= PHYSTOP) || (pa < end)) {
+  if ((pa >= PHYSTOP) || (pa < (uint64)end)) {
     return -1;
   }
   tmp = mem_map[IDX_MEM_MAP(pa)].refCnt;
